Give version resolution a core owner
Move package version resolution into core while keeping the root version module as a compatibility shim, so the root source allow-list contains only entrypoints and shims.

Constraint: src/version had a small production caller surface, while src/mcp-server and src/session-manager remain heavily used compatibility re-exports.

Rejected: moving mcp-server or session-manager root shims in this slice | their broad compatibility surface makes that a separate migration, not a cleanup prerequisite.

Confidence: high

Scope-risk: narrow

Directive: Keep new runtime implementation out of src root; use src/core/version for version resolution and leave root shims thin.

Tested: npm test -- --runTestsByPath tests/cli/doctor.test.ts tests/cli/doctor/runtime-diagnostics.test.ts tests/tools/oc-doctor-report.test.ts tests/mcp/server.test.ts tests/core/metrics/tool-output-observability.test.ts tests/core/skill-memory/oc-skill-tools.test.ts tests/tools/oc-skill-replay.test.ts tests/transports/duplicate-controller-error-server.test.ts --runInBand; npm run build; npm run lint; npm run lint:tier; npm run lint:repo-structure; git diff --check
