+++ src/broker/discovery.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs';
 import * as os from 'os';
 import * as path from 'path';
-import { getVersion } from '../version';
+import { getVersion } from '../core/version';
 import { controllerLockKey, isPidAlive, normalizeControllerUserDataDir } from '../chrome/controller-lock';
 
 export interface BrokerIdentity {
