+++ src/chrome/controller-lock.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs';
 import * as os from 'os';
 import * as path from 'path';
-import { getVersion } from '../version';
+import { getVersion } from '../core/version';
 import { fetchJsonVersion } from './devtools-info';
 import { DEFAULT_CHROME_LAUNCH_TIMEOUT_MS } from '../config/defaults';
