+++ src/index.ts
@@ -22,7 +22,7 @@ import { writePidFile, cleanOrphanedChromeProcesses } from './chrome/pid-manager
 import { installParentWatcher, ParentWatcherHandle } from './core/process/parent-watcher';
 import { installIdleTimeout, IdleTimeoutHandle, parseDuration } from './core/process/idle-timeout';
 import { getIdleState } from './core/activity/idle-state';
-import { getVersion } from './version';
+import { getVersion } from './core/version';
 import { bootstrapPilot, logActiveFlags } from './harness/flags';
 import { ChromeProcessWatchdog } from './chrome/process-watchdog';
 import {
