+++ src/mcp/server.ts
@@ -53,7 +53,7 @@ import { isClientDisconnect } from '../errors/abort';
 import { setLogSender, type LogLevel, logLevelSetErrorOrNull } from '../utils/log';
 import { isAllowed, requiredScope } from '../auth/scope-policy';
 import type { Principal } from '../auth/api-key-types';
-import { getVersion } from '../version';
+import { getVersion } from '../core/version';
 import { isTimeoutError } from '../errors/timeout';
 import { OpenChromeConnectionError } from '../errors/connection';
 import { getTaskJournal } from '../journal/task-journal';
